HDFS-16065. Create RouterClientMetrics for new metrics
